refactor: move LiveSourceLocation
