+++ platform/core/src/main/kotlin/spp/platform/core/SourceService.kt
@@ -47,6 +47,7 @@ import spp.protocol.artifact.ArtifactType
 import spp.protocol.artifact.metrics.MetricStep
 import spp.protocol.instrument.*
 import spp.protocol.instrument.LiveInstrumentType.*
+import spp.protocol.instrument.location.LiveSourceLocation
 import spp.protocol.instrument.meter.MeterType
 import spp.protocol.instrument.meter.MetricValue
 import spp.protocol.instrument.meter.MetricValueType
