+++ platform/core/src/test/kotlin/integration/MetaTest.kt
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 import java.util.concurrent.TimeUnit
 
 @ExtendWith(VertxExtension::class)
