+++ platform/core/src/test/kotlin/integration/VariableControlPermissionTest.kt
@@ -22,7 +22,7 @@ import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 import spp.protocol.instrument.variable.LiveVariableControl
 import spp.protocol.marshall.ServiceExceptionConverter
 import spp.protocol.platform.auth.DeveloperRole
