+++ platform/processor/live-instrument/src/test/kotlin/integration/NegativePrimitiveConditionITTest.kt
@@ -23,7 +23,7 @@ import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 
 @Suppress("UNUSED_VARIABLE", "unused")
 class NegativePrimitiveConditionITTest : LiveInstrumentIntegrationTest() {
