+++ platform/processor/live-instrument/src/test/kotlin/integration/breakpoint/LargeArrayLiveBreakpointTest.kt
@@ -26,7 +26,7 @@ import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 import java.util.*
 
 class LargeArrayLiveBreakpointTest : LiveInstrumentIntegrationTest() {
