+++ platform/processor/live-instrument/src/test/kotlin/integration/breakpoint/LargeMapLiveBreakpointTest.kt
@@ -25,7 +25,7 @@ import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 
 class LargeMapLiveBreakpointTest : LiveInstrumentIntegrationTest() {
