+++ platform/processor/live-instrument/src/test/kotlin/integration/breakpoint/LargeObjectLiveBreakpointTest.kt
@@ -26,7 +26,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 
 @Suppress("UNUSED_VARIABLE")
 class LargeObjectLiveBreakpointTest : LiveInstrumentIntegrationTest() {
