+++ platform/processor/live-instrument/src/test/kotlin/integration/breakpoint/MetaLiveBreakpointTest.kt
@@ -24,7 +24,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 
 class MetaLiveBreakpointTest : LiveInstrumentIntegrationTest() {
