+++ platform/processor/live-instrument/src/test/kotlin/integration/breakpoint/RemoveByLocationLiveBreakpointTest.kt
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveBreakpoint
 import spp.protocol.instrument.LiveInstrumentType
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
