+++ platform/processor/live-instrument/src/test/kotlin/integration/breakpoint/SimpleCollectionsLiveBreakpointTest.kt
@@ -27,7 +27,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 
 @Suppress("UNUSED_VARIABLE")
 class SimpleCollectionsLiveBreakpointTest : LiveInstrumentIntegrationTest() {
