+++ platform/processor/live-instrument/src/test/kotlin/integration/breakpoint/SmallMapLiveBreakpointTest.kt
@@ -26,7 +26,7 @@ import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 
 class SmallMapLiveBreakpointTest : LiveInstrumentIntegrationTest() {
