+++ platform/processor/live-instrument/src/test/kotlin/integration/log/FormatLiveLogTest.kt
@@ -24,7 +24,7 @@ import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import spp.protocol.instrument.LiveLog
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 
 @Suppress("UNUSED_VARIABLE")
 class FormatLiveLogTest : LiveInstrumentIntegrationTest() {
