+++ platform/processor/live-view/src/main/kotlin/spp/processor/live/impl/view/LiveLogView.kt
@@ -34,7 +34,7 @@ import spp.processor.live.impl.view.model.LiveGaugeValueMetrics
 import spp.processor.live.impl.view.util.MetricTypeSubscriptionCache
 import spp.processor.live.impl.view.util.ViewSubscriber
 import spp.protocol.artifact.log.Log
-import spp.protocol.instrument.LiveSourceLocation
+import spp.protocol.instrument.location.LiveSourceLocation
 import java.time.Instant
 import java.time.ZoneOffset
 import java.time.format.DateTimeFormatterBuilder
